requirements.txt
langchain-core
langchain-community
langchain-google-genai
langchain-chroma
chromadb
streamlit
pypdf
python-dotenv